pyproject.toml
[project]
name = "aish"
version = "0.1.0"
description = "AI Shell - A shell with built-in LLM capabilities"
readme = "README.md"
authors = [{ name = "Sian Cao", email = "yinshuiboy@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.9.0",
    "httpx[socks]>=0.24.0",
    "langfuse==2.60.3",
    "litellm==1.74.3",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "tiktoken>=0.5.0",
    "typer>=0.16.0",
    "watchfiles>=0.22.0",
]

[project.scripts]
aish = "aish.cli:main"
aish-sandbox = "aish.sandboxd:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aish"]
include = [
    "src/aish/prompts/*.md",
    "src/aish/i18n/*.yaml",
]
# 包含文档文件到 wheel 包
artifacts = ["docs/skills-guide.md"]
# 将文档安装到系统目录
[tool.hatch.build.targets.wheel.shared-data]
"docs" = "share/doc/aish"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pyinstaller>=6.14.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
# Limit test discovery to this project's tests. The repository also contains
# packaging artifacts under `debian/` which include vendored third-party
# libraries (e.g. jsonschema) with their own test suites that require extra
# optional deps and data files.
testpaths = ["tests"]
pythonpath = ["src"]
norecursedirs = [
    ".git",
    ".venv",
    "build",
    "dist",
    "debian",
]
markers = [
    "asyncio: tests using pytest-asyncio",
    "anyio: tests using pytest-anyio",
]
